Extract `VersionNumber` to standalone artifact
