+++ tools/version-number/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=version-number
+POM_NAME=VersionNumber
+POM_DESCRIPTION=VersionNumber
